breaking: configurable file postfixes for PDS metadata extraction (#124)
* feat: configurable file postfixes for PDS metadata extraction

* fix: some bugfix + update manual deployment update

* breaking: rename lvl0 and lvl1 to s4pa and cas

* fix: rename output variables

* chore: update contributing
