+++ cumulus_lambda_functions/metadata_cas_generate_cmr/generate_cmr.py
@@ -8,8 +8,8 @@ from cumulus_lambda_functions.lib.aws.aws_s3 import AwsS3
 from cumulus_lambda_functions.lib.json_validator import JsonValidator
 from cumulus_lambda_functions.lib.lambda_logger_generator import LambdaLoggerGenerator
 from cumulus_lambda_functions.lib.time_utils import TimeUtils
-from cumulus_lambda_functions.snpp_level1a_generate_cmr.echo_metadata import EchoMetadata
-from cumulus_lambda_functions.snpp_level1a_generate_cmr.l1a_input_metadata import L1AInputMetadata
+from cumulus_lambda_functions.metadata_cas_generate_cmr.echo_metadata import EchoMetadata
+from cumulus_lambda_functions.metadata_cas_generate_cmr.l1a_input_metadata import L1AInputMetadata
 
 LOGGER = LambdaLoggerGenerator.get_logger(__name__, LambdaLoggerGenerator.get_level_from_env())
