+++ cumulus_lambda_functions/metadata_cas_generate_cmr/l1a_input_metadata.py
