+++ tf-module/unity-cumulus/outputs.tf
@@ -1,5 +1,5 @@
-output "snpp_lvl0_generate_cmr_arn" {
-  value = aws_lambda_function.snpp_lvl0_generate_cmr.arn
+output "metadata_s4pa_generate_cmr_arn" {
+  value = aws_lambda_function.metadata_s4pa_generate_cmr.arn
 }
 
 output "cumulus_granules_dapa_arn" {
@@ -10,8 +10,8 @@ output "cumulus_collections_dapa_arn" {
   value = aws_lambda_function.cumulus_collections_dapa.arn
 }
 
-output "snpp_lvl1_generate_cmr_arn" {
-    value = aws_lambda_function.snpp_lvl1_generate_cmr.arn
+output "metadata_cas_generate_cmr_arn" {
+    value = aws_lambda_function.metadata_cas_generate_cmr.arn
 }
 
 output "cumulus_collections_ingest_cnm_dapa_arn" {
